ECI/src/requirements.txt
flask
python-dotenv
psycopg2-binary
python-decouple
flask_sqlalchemy
flask-CORS
flask_migrate
mysqlclient
requests
PyJWT 
Flask-JWT-Extended